requirements.txt
pymongo
dnspython
python-dotenv
playwright
PyMuPDF
